+++ src/main/java/org/ballerinalang/sql/utils/QueryUtils.java
@@ -20,15 +20,10 @@ package org.ballerinalang.sql.utils;
 import org.ballerinalang.jvm.BallerinaValues;
 import org.ballerinalang.jvm.scheduling.Scheduler;
 import org.ballerinalang.jvm.scheduling.Strand;
-import org.ballerinalang.jvm.types.BArrayType;
 import org.ballerinalang.jvm.types.BField;
-import org.ballerinalang.jvm.types.BPackage;
 import org.ballerinalang.jvm.types.BRecordType;
 import org.ballerinalang.jvm.types.BStreamType;
 import org.ballerinalang.jvm.types.BStructureType;
-import org.ballerinalang.jvm.types.BType;
-import org.ballerinalang.jvm.types.BTypes;
-import org.ballerinalang.jvm.types.TypeFlags;
 import org.ballerinalang.jvm.util.Flags;
 import org.ballerinalang.jvm.values.AbstractObjectValue;
 import org.ballerinalang.jvm.values.ErrorValue;
@@ -44,17 +39,10 @@ import org.ballerinalang.sql.exception.ApplicationError;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
-import java.sql.Statement;
-import java.sql.Types;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
-import java.util.Set;
 
 /**
  * This class provides the util implementation which executes sql queries.
@@ -88,8 +76,8 @@ public class QueryUtils {
                 List<ColumnDefinition> columnDefinitions;
                 BStructureType streamConstraint;
                 if (recordType == null) {
-                    columnDefinitions = getColumnDefinitions(resultSet, null);
-                    BRecordType defaultRecord = getDefaultStreamConstraint();
+                    columnDefinitions = Utils.getColumnDefinitions(resultSet, null);
+                    BRecordType defaultRecord = Utils.getDefaultStreamConstraint();
                     Map<String, BField> fieldMap = new HashMap<>();
                     for (ColumnDefinition column : columnDefinitions) {
                         int flags = Flags.PUBLIC;
@@ -105,15 +93,15 @@ public class QueryUtils {
                     streamConstraint = defaultRecord;
                 } else {
                     streamConstraint = (BStructureType) ((TypedescValue) recordType).getDescribingType();
-                    columnDefinitions = getColumnDefinitions(resultSet, streamConstraint);
+                    columnDefinitions = Utils.getColumnDefinitions(resultSet, streamConstraint);
                 }
-                return new StreamValue(new BStreamType(streamConstraint), createRecordIterator(resultSet,
+                return new StreamValue(new BStreamType(streamConstraint), Utils.createRecordIterator(resultSet,
                         statement, connection, columnDefinitions, streamConstraint));
             } catch (SQLException e) {
                 Utils.closeResources(strand, resultSet, statement, connection);
                 ErrorValue errorValue = ErrorGenerator.getSQLDatabaseError(e,
                         "Error while executing sql query: " + sqlQuery + ". ");
-                return new StreamValue(new BStreamType(getDefaultStreamConstraint()), createRecordIterator(errorValue));
+                return new StreamValue(new BStreamType(Utils.getDefaultStreamConstraint()), createRecordIterator(errorValue));
             } catch (ApplicationError applicationError) {
                 Utils.closeResources(strand, resultSet, statement, connection);
                 ErrorValue errorValue = ErrorGenerator.getSQLApplicationError(applicationError.getMessage());
@@ -137,86 +125,11 @@ public class QueryUtils {
 
     private static StreamValue getErrorStream(Object recordType, ErrorValue errorValue) {
         if (recordType == null) {
-            return new StreamValue(new BStreamType(getDefaultStreamConstraint()), createRecordIterator(errorValue));
+            return new StreamValue(new BStreamType(Utils.getDefaultStreamConstraint()), createRecordIterator(errorValue));
         } else {
             return new StreamValue(new BStreamType(((TypedescValue) recordType).getDescribingType()),
                     createRecordIterator(errorValue));
         }
-    }
-
-    private static BRecordType getDefaultStreamConstraint() {
-        BRecordType defaultRecord = new BRecordType("$stream$anon$constraint$",
-                new BPackage("ballerina", "lang.annotations", "0.0.0"), 0, false,
-                TypeFlags.asMask(TypeFlags.ANYDATA, TypeFlags.PURETYPE));
-        defaultRecord.restFieldType = BTypes.typeAnydata;
-        return defaultRecord;
-    }
-
-    private static List<ColumnDefinition> getColumnDefinitions(ResultSet resultSet, BStructureType streamConstraint)
-            throws SQLException, ApplicationError {
-        List<ColumnDefinition> columnDefs = new ArrayList<>();
-        Set<String> columnNames = new HashSet<>();
-        ResultSetMetaData rsMetaData = resultSet.getMetaData();
-        int cols = rsMetaData.getColumnCount();
-        for (int i = 1; i <= cols; i++) {
-            String colName = rsMetaData.getColumnLabel(i);
-            if (columnNames.contains(colName)) {
-                String tableName = rsMetaData.getTableName(i).toUpperCase(Locale.getDefault());
-                colName = tableName + "." + colName;
-            }
-            int sqlType = rsMetaData.getColumnType(i);
-            String sqlTypeName = rsMetaData.getColumnTypeName(i);
-            boolean isNullable = true;
-            if (rsMetaData.isNullable(i) == ResultSetMetaData.columnNoNulls) {
-                isNullable = false;
-            }
-            columnDefs.add(generateColumnDefinition(colName, sqlType, sqlTypeName, streamConstraint, isNullable));
-            columnNames.add(colName);
-        }
-        return columnDefs;
-    }
-
-    private static ColumnDefinition generateColumnDefinition(String columnName, int sqlType, String sqlTypeName,
-                                                             BStructureType streamConstraint, boolean isNullable)
-            throws ApplicationError {
-        String ballerinaFieldName = null;
-        BType ballerinaType = null;
-        if (streamConstraint != null) {
-            for (Map.Entry<String, BField> field : streamConstraint.getFields().entrySet()) {
-                if (field.getKey().equalsIgnoreCase(columnName)) {
-                    ballerinaFieldName = field.getKey();
-                    ballerinaType = Utils.validFieldConstraint(sqlType, field.getValue().type);
-                    if (ballerinaType == null) {
-                        throw new ApplicationError(field.getValue().type.getName() + " cannot be mapped to sql type '"
-                                + sqlTypeName + "'");
-                    }
-                    break;
-                }
-            }
-            if (ballerinaFieldName == null) {
-                throw new ApplicationError("No mapping field found for SQL table column '" + columnName + "'"
-                        + " in the record type '" + streamConstraint.getName() + "'");
-            }
-        } else {
-            ballerinaType = getDefaultBallerinaType(sqlType);
-            ballerinaFieldName = columnName;
-        }
-        return new ColumnDefinition(columnName, ballerinaFieldName, sqlType, sqlTypeName, ballerinaType, isNullable);
-
-    }
-
-    private static ObjectValue createRecordIterator(ResultSet resultSet,
-                                                    Statement statement,
-                                                    Connection connection, List<ColumnDefinition> columnDefinitions,
-                                                    BStructureType streamConstraint) {
-        ObjectValue resultIterator = BallerinaValues.createObjectValue(Constants.SQL_PACKAGE_ID,
-                Constants.RESULT_ITERATOR_OBJECT, new Object[1]);
-        resultIterator.addNativeData(Constants.RESULT_SET_NATIVE_DATA_FIELD, resultSet);
-        resultIterator.addNativeData(Constants.STATEMENT_NATIVE_DATA_FIELD, statement);
-        resultIterator.addNativeData(Constants.CONNECTION_NATIVE_DATA_FIELD, connection);
-        resultIterator.addNativeData(Constants.COLUMN_DEFINITIONS_DATA_FIELD, columnDefinitions);
-        resultIterator.addNativeData(Constants.RECORD_TYPE_DATA_FIELD, streamConstraint);
-        return resultIterator;
     }
 
     private static ObjectValue createRecordIterator(ErrorValue errorValue) {
@@ -224,52 +137,4 @@ public class QueryUtils {
                 errorValue);
     }
 
-    private static BType getDefaultBallerinaType(int sqlType) {
-        switch (sqlType) {
-            case Types.ARRAY:
-                return new BArrayType(BTypes.typeAnydata);
-            case Types.CHAR:
-            case Types.VARCHAR:
-            case Types.LONGVARCHAR:
-            case Types.NCHAR:
-            case Types.NVARCHAR:
-            case Types.LONGNVARCHAR:
-            case Types.CLOB:
-            case Types.NCLOB:
-            case Types.DATE:
-            case Types.TIME:
-            case Types.TIMESTAMP:
-            case Types.TIMESTAMP_WITH_TIMEZONE:
-            case Types.TIME_WITH_TIMEZONE:
-                return BTypes.typeString;
-            case Types.TINYINT:
-            case Types.SMALLINT:
-            case Types.INTEGER:
-            case Types.BIGINT:
-                return BTypes.typeInt;
-            case Types.BIT:
-            case Types.BOOLEAN:
-                return BTypes.typeBoolean;
-            case Types.NUMERIC:
-            case Types.DECIMAL:
-                return BTypes.typeDecimal;
-            case Types.REAL:
-            case Types.FLOAT:
-            case Types.DOUBLE:
-                return BTypes.typeFloat;
-            case Types.BLOB:
-            case Types.BINARY:
-            case Types.VARBINARY:
-            case Types.LONGVARBINARY:
-            case Types.ROWID:
-                return new BArrayType(BTypes.typeByte);
-            case Types.REF:
-            case Types.STRUCT:
-                return getDefaultStreamConstraint();
-            case Types.SQLXML:
-                return BTypes.typeXML;
-            default:
-                return BTypes.typeAnydata;
-        }
-    }
 }
